% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitting_algorithm.R
\name{get_mean_compare}
\alias{get_mean_compare}
\title{plot function to compare the estimation of SFPDM mean function and true mean function}
\usage{
get_mean_compare(mean_func, Spline_func, theta_est, othermodel = NULL)
}
\arguments{
\item{mean_func}{the true mean function}

\item{Spline_func}{spline function used to trrain the model}

\item{theta_est}{the estimation of theta parameters from the results of train_function}

\item{othermodel}{othermodel if users want to add to compare}
}
\value{
the plot of comparison between the estimation of SFPDM mean function and true mean function
}
\description{
plot function to compare the estimation of SFPDM mean function and true mean function
}
\examples{
tmin = 0 # the start point of the curve
tmax = 1 # the end point of the curve, i.e. the cuvre's regin is from tmin to tmax
ymin = 0 # the minimum of the covariates
ymax = 1 # the maximum of the covariate
num_bin = 20 # number of bin in initial steps of our algorithm
order = 4 # spline order
nknots =8 # number of knots
splineObj_t = new(orthoSpline,tmin,tmax,order,nknots)
# degree freedom of spline basis
M1 = splineObj_t$getDoF()
## basis with respect to y to get the tensor product basis
yknots = 3
splineObj_y = new(orthoSpline,ymin,ymax,order,(yknots))
# degree freedom of spline basis
M2 = splineObj_y$getDoF()
## basis with d that is in matrix C
# basis with respect y (d^T), in C matrix
dknots = 5
splineObj_d = new(orthoSpline,ymin,ymax,order,dknots)
# degree freedom of spline basis
M3 = splineObj_d$getDoF()
k = M3
Spline_func = c(splineObj_t,splineObj_d,splineObj_y)
## set the number of principle components
Eig_num = r =  3
## the the number of spline basis used
## spline basis for t
M1 = 10
M2 = 5
M3 = k = 7
theta = rep(0,M1*M2)
sigma2 = 0.01
beta = rep(0,r*M1*M3)
parameter_best = train_function (Data_generated = Data_generated,Eig_num = Eig_num,k = k, beta = beta,theta = theta,sigma2 = sigma2)
mean_func = function(t,covariate){return(30*(t - covariate)^2)}
get_mean_compare(mean_func = mean_func,Spline_func = Spline_func,theta_est = parameter_best$theta)
}
